refactor: SubwayTestUtil 위치 수정
utils 패키지에 있는 것이 더 자연스러워 위치를 수정합니다.
